update remote and messaging sample to align with base
* Added server.js to run app from xmpls
* cleaned up package.json and bumped versions
