+++ xmpls/messaging/app/incidents/field.cds
@@ -1,5 +1,4 @@
-using ProcessorService as service from '../../srv/services.cds';
-using from './annotations.cds';
+using ProcessorService as service from '../../services';
 annotate service.Incidents with @(
     /*adding email to the object page enables users to view the
     changes recieved via Messaging/Eventing*/
