+++ xmpls/messaging/package.json
@@ -2,84 +2,29 @@
     "name": "@capire/messaging",
     "version": "1.0.0",
     "dependencies": {
-        "@sap-cloud-sdk/connectivity": "^3.6.0",
-        "@sap-cloud-sdk/http-client": "^3.6.0",
-        "@sap-cloud-sdk/resilience": "^3.6.0",
-        "@sap-cloud-sdk/util": "^3.6.0",
-        "@sap/cds": ">=7",
-        "@cap-js/hana": ">=1",
-        "@sap/xssec": "^3",
+        "@sap-cloud-sdk/connectivity": "4.0.2",
+        "@sap-cloud-sdk/http-client": "4.0.2",
+        "@sap-cloud-sdk/resilience": "4.0.2",
+        "@sap-cloud-sdk/util": "4.0.2",
+        "@sap/cds": ">=8",
         "express": "^4",
-        "passport": "^0",
-        "@sap/xb-msg-amqp-v100": "^0"
+        "@sap/xb-msg-amqp-v100": "^0.9.58"
     },
     "devDependencies": {
         "@cap-js/cds-test": "*",
-        "@cap-js/sqlite": "*",
-        "@sap/cds-dk": "*",
-        "@sap/ux-specification": "^1.108.4"
+        "@cap-js/sqlite": "*"
     },
     "scripts": {
         "watch": "cds watch",
         "start": "cds-serve",
-        "test": "npx jest --silent",
-        "add-change-tracking": "npm add @cap-js/change-tracking && cp xmpls/change-tracking.cds ./srv && cp xmpls/change-tracking.test.js ./test",
-        "add-telemetry": "npm add @cap-js/telemetry",
-        "add-attachments": "npm add @cap-js/attachments && cp xmpls/attachments.cds ./srv && cp -r xmpls/content ./db/data/content",
-        "clone-add-attachments": "git clone https://github.com/cap-js/attachments.git && cp -r attachments/xmpl/db . && cp -r attachments/xmpl/app . && npm add https://github.com/cap-js/attachments.git",
-        "add-notifications": "npm add @cap-js/notifications && cp xmpls/alert-notifications.js ./srv && cp xmpls/notification-types.json ./srv",
-        "add-audit-log": "npm add @cap-js/audit-logging && cp xmpls/data-privacy.cds ./srv && cp xmpls/audit-log.test.js ./test",
-        "add-remote-service": "cp -r xmpls/remote-service/* .",
-        "add-messaging": "cp -r xmpls/messaging/* .",
-        "add-all-xmpls": "npm run add-remote-service && npm run add-messaging && npm run add-change-tracking && npm run add-audit-log",
-        "reset": "read -p 'This will irreversibly reset your working directory including ALL files in this git repo. Continue?' -n 1 -r && echo && if [[ $REPLY =~ ^[Yy]$ ]]; then git clean -fd && git reset --hard && npm i; fi"
+        "test": "npx jest --silent"
     },
     "private": true,
     "sapux": [
         "app/incidents"
     ],
-    "jest": {
-        "modulePathIgnorePatterns": [
-            "<rootDir>/xmpls/"
-        ]
-    },
     "cds": {
         "requires": {
-            "incidents-app": {
-                "model": "@capire/incidents"
-            },
-            "[production]": {
-                "db": {
-                    "kind": "hana-cloud"
-                },
-                "auth": {
-                    "kind": "xsuaa"
-                },
-                "approuter": {
-                    "kind": "cloudfoundry"
-                }
-            },
-            "[development]": {
-                "auth": {
-                    "kind": "mocked",
-                    "users": {
-                        "alice": {
-                            "roles": [
-                                "support",
-                                "admin"
-                            ]
-                        },
-                        "bob": {
-                            "roles": [
-                                "support"
-                            ]
-                        }
-                    }
-                }
-            },
-            "db": {
-                "kind": "sql"
-            },
             "API_BUSINESS_PARTNER": {
                 "kind": "odata-v2",
                 "model": "srv/external/API_BUSINESS_PARTNER",
