+++ xmpls/messaging/services.cds
@@ -1,3 +1,4 @@
 // Using local service implementation
 using { ProcessorService } from '../../app/services';
-annotate ProcessorService with @impl: 'srv/services.js';
+annotate ProcessorService with @impl: 'srv/processor-service.js';
+using from './app/incidents/field';
