+++ xmpls/remote-service/services.cds
@@ -1,3 +1,3 @@
 // Using local service implementation
 using { ProcessorService } from '../../app/services';
-annotate ProcessorService with @impl: 'srv/services.js';
+annotate ProcessorService with @impl: 'srv/processor-service.js';
